Extract editor selection debug helpers
