+++ src/game_data_internal.h
@@ -1051,6 +1051,10 @@ slayer3d_game_data_sector_level_variant sector_level_variant_from_string(const c
                                                                          bool sector_lighting_enabled);
 slayer3d_bounding_box translated_bounds(slayer3d_bounding_box bounds, slayer3d_vec3 position);
 bool model_bounds(const slayer3d_model *model, slayer3d_bounding_box *out_bounds);
+void init_editor_selection(slayer3d_game_data_editor_selection *selection);
+bool editor_selection_active_for_scene(const slayer3d_game_data_runtime *runtime);
+bool editor_command_preview_active_for_scene(const slayer3d_game_data_runtime *runtime);
+bool editor_placement_preview_active_for_scene(const slayer3d_game_data_runtime *runtime);
 bool compile_brush_world_visibility_grid(brush_world_runtime *world_runtime);
 void free_brush_world_visibility_grid(brush_world_runtime *world_runtime);
 bool rebuild_brush_world_runtime_artifacts(brush_world_runtime *world_runtime, char *error_buffer,
